Refactoring Process Phase 1.3

+++ scrapapp.webui/Models/OrderListModel.cs
@@ -18,7 +18,7 @@ namespace scrapapp.webui.Models
         public EnumPaymentType PaymentType { get; set; }
         public EnumOrderState OrderState { get; set; }
         public List<OrderItemModel> OrderItems { get; set; }
-        public double TotalPrice()
+        public decimal TotalPrice()
         {
             return OrderItems.Sum(i => i.Price * i.Quantity);
         }
@@ -27,7 +27,7 @@ namespace scrapapp.webui.Models
     public class OrderItemModel
     {
         public int OrderItemId { get; set; }
-        public double Price { get; set; }
+        public decimal Price { get; set; }
         public string Name { get; set; }
         public string ImageUrl { get; set; }
         public int Quantity { get; set; }
